Use size_t for length in replaycache_add_test_and_elapsed()
For consistency with the rest of buffer lengths in Tor.

See #8960

+++ src/or/rendservice.c
@@ -1207,7 +1207,7 @@ rend_service_introduce(origin_circuit_t *circuit, const uint8_t *request,
   /* check for replay of PK-encrypted portion. */
   replay = replaycache_add_test_and_elapsed(
     intro_point->accepted_intro_rsa_parts,
-    parsed_req->ciphertext, (int)parsed_req->ciphertext_len,
+    parsed_req->ciphertext, parsed_req->ciphertext_len,
     &elapsed);
 
   if (replay) {
